fix: remaining test contract fixes and docs expansion
- temporal smoother: clamp only when blending (alpha < 1.0)
- depth curve: sort before validating; linear 2-point curve uses np.interp
- batch queue: cycle detection via dependents with direct check; skip completed jobs
- thumbnails: deduplicate frame indices when count exceeds frame range
- video writer: validate dimensions/fps at construction
- esrgan/spatial/audio/auth: lazy imports and contract fixes
- developer guide: expand prose to 3000+ words, 5+ bash examples

+++ src/video2d3d/skybox/processor.py
@@ -397,7 +397,7 @@ def create_sky_depth_mask(
         y_coords = np.arange(h).reshape(-1, 1)
         normalized_y = np.clip(y_coords / horizon_y, 0, 1)
         gradient_depth = max_depth * (1 - gradient_strength * normalized_y)
-        depth_mask = np.where(sky_mask, gradient_depth, max_depth)
+        depth_mask = np.where(sky_mask, gradient_depth, max_depth).astype(np.float32)
 
     return depth_mask
